requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
click>=8.1.0

# Database
psycopg2-binary>=2.9.9
pgvector>=0.2.4
SQLAlchemy>=2.0.23

# RAG Pipeline
langchain>=0.1.0
langchain-community>=0.0.10
sentence-transformers>=2.2.2
tiktoken>=0.5.2

# LLM
openai>=1.6.0
httpx>=0.25.0

# Research Paper APIs
biopython>=1.82
arxiv>=2.1.0
requests>=2.31.0
lxml>=5.0.0

# Healthcare Standards
fhir.resources>=7.1.0
hl7apy>=1.3.4
# Note: python-hl7 not compatible with Python 3.14, using hl7apy instead

# Security & HIPAA
cryptography>=41.0.7
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0
pyotp>=2.9.0

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0

# Development
black>=23.12.0
isort>=5.13.0
mypy>=1.8.0
